+++ aop-navigation/projects/aop-routing/src/lib/navigation/aop-proxy-navigation.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { NavAux } from '../model/models';
+import { NavAux } from '../model/nav-aux';
 
 @Injectable({
   providedIn: 'root',
